requirements.txt
# Telegram Bot
python-telegram-bot>=20.0
python-dotenv>=0.19.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.6

# Google AI
google-generativeai>=0.3.0

# Utils
aiohttp>=3.9.1